gps: move TestBadSolveOpts to solver_inputs_test.go
